requirements.txt
streamlit
requests
beautifulsoup4
pdfplumber
pytesseract
pillow
fuzzywuzzy
python-Levenshtein
pdf2image
